Update patch set 1
Patch Set 1: Code-Review-2

Per-file fsync hacks are horrible for data integrity.  This is one of the categories of stuff that should not be in any omni kernel.

If there are any devices with this, it needs to be ripped out.

Patch-set: 1
Reviewer: Gerrit User 1000008 <1000008@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=-2